Extract strategy for creating data structure for holding evolving particles
The structure for calculating the time evolution now fully stores the state
of the evolution in an implementation of a trait and delegates write
operations to it, and requires a means of creating such implementations as
part of its construction.

There is only one choice for this at the moment: the maximally contiguous
vector of structs.