Add implementation of memory strategy using a vector of boxed structs
Now boxed MassNormalizedWithForceField structs implement the trait, and
correspondingly their vectors the relevant collection trait. The memory
management has been further delegated and the collection does not have
to expose any indices, at the expense of a slightly cumbersome model of
updating pairs with an intermediate new object and a pair of functions
to update each member of the pair sequentially.

There is still no option to select the boxed struct memory model yet.